ensure uri-compatible strings are used on windows
Since the first two arguments to io.bit3.jsass.importer.Import. are
going to be treated as URIs we need to ensure that the strings passed to
the constructor are URI compatible (i.e. not contain backslashes or
other illegal characters).  This motivates both changes.

Also we don't want windows users to have to import webjars using
something like  `@import "boostrap\\scss/bootstrap`.  This is another
reason for the change to webjars.clj.