Fix leiningen infinite loop or crash in prep-tasks
Fixes #18

When a user runs the sass4clj lein task, lein also runs prep-tasks. If
the sass4clj task is one of the prep-tasks, a new instance of the task
is run, eventually resulting in a stack overflow.

None of the prep-tasks are actually needed to successfully run the
task. This commit ignores the prep-tasks when the sass4clj lein task
is run.

See also: https://github.com/emezeske/lein-cljsbuild/issues/451